Initial implementation completed
Completed the implementation that counts number of measurements that were bigger than the previous one. Just a simple loop.

Removed the unit test that tested for an empty array. It returned a compile error because [] is technically of type [Any]. Did not want to bother since it was an edge case anyway.

Fixed wrong spelling on calls to `increases()` in unit tests.