Part 1 of Day 1 puzzle completed, answer is 1121
Completed the first half of the day 1 challenge. The input was considerably longer than I anticipated. Saved it to the file input.txt and added file reading logic in Julia. Was pleasantly surprised how easy and neat the code was. Remind me a lot of writing in Python but with the option to specify the types for clarity.

The last expression in a function is returned so there is no need to use return.